#24 add_init_page
